+++ sandpack-client/src/client.ts
@@ -1,6 +1,5 @@
 import { getTemplate } from "codesandbox-import-utils/lib/create-sandbox/templates";
 import isEqual from "lodash.isequal";
-import { SandpackLogLevel } from ".";
 
 import Protocol from "./file-resolver-protocol";
 import { IFrameProtocol } from "./iframe-protocol";
@@ -21,6 +20,8 @@ import {
   addPackageJSONIfNeeded,
   extractErrorDetails,
 } from "./utils";
+
+import { SandpackLogLevel } from ".";
 
 export interface ClientOptions {
   /**
@@ -156,9 +157,12 @@ export class SandpackClient {
       );
     }
 
-    this.iframe.src = options.startRoute
+    const urlSource = options.startRoute
       ? new URL(options.startRoute, this.bundlerURL).toString()
       : this.bundlerURL;
+
+    this.iframe.contentWindow?.location.replace(urlSource);
+
     this.iframeProtocol = new IFrameProtocol(this.iframe, this.bundlerURL);
 
     this.unsubscribeGlobalListener = this.iframeProtocol.globalListen(
@@ -233,7 +237,10 @@ export class SandpackClient {
     sandboxInfo = this.sandboxInfo,
     isInitializationCompile?: boolean
   ): void {
-    this.sandboxInfo = sandboxInfo;
+    this.sandboxInfo = {
+      ...this.sandboxInfo,
+      ...sandboxInfo,
+    };
 
     const files = this.getFiles();
 
@@ -294,7 +301,7 @@ export class SandpackClient {
       showLoadingScreen: this.options.showLoadingScreen ?? true,
       skipEval: this.options.skipEval || false,
       clearConsoleDisabled: !this.options.clearConsoleOnFirstCompile,
-      logLevel: this.options.logLevel,
+      logLevel: this.options.logLevel ?? SandpackLogLevel.Info,
     });
   }