+++ sandpack-react/.storybook/preview.js
@@ -2,4 +2,9 @@ import "../src/styles/index.css";
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
+  options: {
+    storySort: {
+      order: ["Intro"],
+    },
+  },
 };