+++ sandpack-react/src/common/OpenInCodeSandboxButton/UnstyledOpenInCodeSandboxButton.tsx
@@ -1,9 +1,17 @@
+/* eslint-disable @typescript-eslint/explicit-function-return-type */
 import type { SandpackBundlerFiles } from "@codesandbox/sandpack-client";
-import { getParameters } from "codesandbox-import-utils/lib/api/define";
+import LZString from "lz-string";
 import * as React from "react";
 
 import { useSandpack } from "../../hooks/useSandpack";
 import type { SandboxEnvironment } from "../../types";
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const getParameters = (parameters: Record<string, any>): string =>
+  LZString.compressToBase64(JSON.stringify(parameters))
+    .replace(/\+/g, "-") // Convert '+' to '-'
+    .replace(/\//g, "_") // Convert '/' to '_'
+    .replace(/=+$/, ""); /* Remove ending '='*/
 
 const CSB_URL = "https://codesandbox.io/api/v1/sandboxes/define";
 
@@ -66,7 +74,9 @@ export const UnstyledOpenInCodeSandboxButton: React.FC<
     [sandpack.activePath, sandpack.environment, sandpack.files]
   );
 
-  // Register the usage of the codesandbox link
+  /**
+   * Register the usage of the codesandbox link
+   */
   React.useEffect(function registerUsage() {
     sandpack.openInCSBRegisteredRef.current = true;
     // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -86,8 +96,8 @@ export const UnstyledOpenInCodeSandboxButton: React.FC<
         <form ref={formRef} action={CSB_URL} method="POST" target="_blank">
           {Array.from(
             paramsValues as unknown as Array<[string, string]>,
-            ([k, v]) => (
-              <input key={k} name={k} type="hidden" value={v} />
+            ([key, value]) => (
+              <input key={key} name={key} type="hidden" value={value} />
             )
           )}
         </form>