+++ sandpack-react/src/components/FileExplorer/FileExplorer.stories.tsx
@@ -4,7 +4,10 @@ import { SandpackLayout } from "../../common/Layout";
 import { SandpackCodeEditor } from "../../components/CodeEditor";
 import { SandpackProvider } from "../../contexts/sandpackContext";
 
-import { FileExplorer } from "./index";
+import { Directory } from "./Directory";
+import { File } from "./File";
+
+import { FileExplorer } from "./";
 
 export default {
   title: "components/File Explorer",
@@ -27,6 +30,7 @@ export const Component: React.FC = () => (
         <SandpackCodeEditor />
       </SandpackLayout>
     </SandpackProvider>
+
     <SandpackProvider
       customSetup={{
         entry: "/index.tsx",
@@ -43,4 +47,44 @@ export const Component: React.FC = () => (
       </SandpackLayout>
     </SandpackProvider>
   </>
+);
+
+export const LongFileTree: React.FC = () => (
+  <SandpackProvider
+    customSetup={{
+      entry: "/index.tsx",
+      files: new Array(20).fill(" ").reduce((acc, _curr, index) => {
+        acc[`/src/com${index}.js`] = "";
+
+        return acc;
+      }, {}),
+    }}
+  >
+    <SandpackLayout>
+      <FileExplorer />
+    </SandpackLayout>
+  </SandpackProvider>
+);
+
+export const FileStory: React.FC = () => (
+  <SandpackProvider>
+    <SandpackLayout>
+      <File depth={1} path="file.ts" />
+    </SandpackLayout>
+  </SandpackProvider>
+);
+
+export const DirectoryIconStory: React.FC = () => (
+  <SandpackProvider>
+    <SandpackLayout>
+      <Directory
+        activePath="file.ts"
+        depth={1}
+        files={{ App: { code: "" } }}
+        prefixedPath="/src"
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        selectFile={(): any => null}
+      />
+    </SandpackLayout>
+  </SandpackProvider>
 );