+++ sandpack-react/src/presets/Sandpack.tsx
@@ -1,4 +1,5 @@
 import { ClasserProvider } from "@code-hike/classer";
+import type { SandpackLogLevel } from "@codesandbox/sandpack-client";
 import * as React from "react";
 
 import { SandpackLayout } from "../common/Layout";
@@ -36,6 +37,7 @@ export interface SandpackProps {
     showNavigator?: boolean;
     showLineNumbers?: boolean;
     showInlineErrors?: boolean;
+    showOpenInCodeSandbox?: boolean;
     showTabs?: boolean;
     closableTabs?: boolean;
     wrapContent?: boolean;
@@ -68,6 +70,7 @@ export interface SandpackProps {
      * appears when `readOnly` is `true`
      */
     showReadOnly?: boolean;
+    logLevel?: SandpackLogLevel;
   };
 }
 
@@ -88,6 +91,7 @@ export const Sandpack: React.FC<SandpackProps> = (props) => {
 
   const previewOptions: PreviewProps = {
     showNavigator: props.options?.showNavigator,
+    showOpenInCodeSandbox: props?.options?.showOpenInCodeSandbox,
   };
 
   const codeEditorOptions: CodeEditorProps = {
@@ -116,6 +120,7 @@ export const Sandpack: React.FC<SandpackProps> = (props) => {
     initMode: props.options?.initMode,
     initModeObserverOptions: props.options?.initModeObserverOptions,
     externalResources: props.options?.externalResources,
+    logLevel: props.options?.logLevel,
   };
 
   // Parts are set as `flex` values, so they set the flex shrink/grow